+++ src/test/java/org/intellij/sdk/language/TableTestParsingTest.java
@@ -13,6 +13,11 @@ public class TableTestParsingTest extends ParsingTestCase {
         doTest(true);
     }
 
+    public void testFirstLineComment() {
+        // Bug: first-line comment is incorrectly parsed as header
+        doTest(true);
+    }
+
     /**
      * @return path to test data file directory relative to root of this module.
      */